+++ examples_and_samples/demo/legacy-old/examples/nuclear-boot-rust/target/doc/bootloader_api/config/sidebar-items.js
@@ -1,1 +1,0 @@
-window.SIDEBAR_ITEMS = {"enum":["Mapping"],"struct":["ApiVersion","BootloaderConfig","FrameBuffer","Mappings"]};